+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Forge
+
+#### Added
+
+- Debugging traces now display the class hash for forked contracts.
+
 ## [0.62.0] - 2026-06-25
 
 ### Forge